kpkg: feat: remove mv and go back to extracting tarballs
This wasn't how i wanted to do this, but the Nim implementation was buggy for no reason.
Even though moveFile should crash if it fails, it succeeded *and* didn't put the file in place.
I am assuming this is a Nim bug and hoping it gets fixed ASAP.
For now this will do.