chkupd: feat: put RELEASE with the Python version on Python dependents
This is done so we avoid module issues with each Python release.
It is hardcoded for now, if other languages follow same kind of path I
may make it into a config.

Example: RELEASE=1 after this update on Meson should become
RELEASE=1-3.13.0

Not an ideal or clean solution, but I'm open for ideas.