kpkg: feat: wip: modularize and simplify builder (also named builder-ng)
It is time for another rewrite/refactor folks!

This time the attention is on the builder, as it has grown too big and too complex.
We will modularize the builder, simplify it, and hopefully introduce some cool new features like proper reproducibility and a more modular runFile syntax, using our own parser.